Downloader API/pyproject.toml
[project]
name = "imdb-dataset-downloader"
version = "1.0.0"
description = "A FastAPI application for downloading and extracting IMDb datasets"
authors = [
    {name = "IMDb Dataset Downloader Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.9.1",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "asyncio==3.4.3",
    "attrs==24.2.0",
    "azure-storage-blob==12.19.0",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.13.3",
    "black==24.1.1",
    "CacheControl==0.14.3",
    "cachetools==5.3.3",
    "certifi==2024.2.2",
    "cffi==1.17.1",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloud-sql-python-connector==1.15.0",
    "coverage==7.4.1",
    "cryptography==44.0.0",
    "db-dtypes==1.3.1",
    "distlib==0.3.8",
    "dnspython==2.7.0",
    "ecdsa==0.19.0",
    "fastapi==0.104.1",
    "filelock==3.16.0",
    "firebase_admin==7.1.0",
    "flake8==7.0.0",
    "frozenlist==1.5.0",
    "google==3.0.0",
    "google-analytics-data==0.18.16",
    "google-api-core==2.27.0",
    "google-api-python-client==2.156.0",
    "google-auth==2.27.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "google-cloud==0.34.0",
    "google-cloud-bigquery==3.17.1",
    "google-cloud-bigquery-datatransfer==3.18.0",
    "google-cloud-bigquery-storage==2.27.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.21.0",
    "google-cloud-secret-manager==2.20.0",
    "google-cloud-storage==3.4.1",
    "google-crc32c==1.6.0",
    "google-oauth2-tool==0.0.3",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.63.0",
    "grpc-google-iam-v1==0.13.0",
    "grpcio==1.64.0",
    "grpcio-status==1.62.2",
    "gspread==6.1.4",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.7",
    "inflect==7.4.0",
    "iniconfig==2.0.0",
    "Mako==1.3.10",
    "MarkupSafe==3.0.3",
    "mccabe==0.7.0",
    "more-itertools==10.5.0",
    "msgpack==1.1.2",
    "multidict==6.1.0",
    "mypy==1.8.0",
    "mypy_extensions==1.1.0",
    "numpy==1.24.3",
    "oauth2client==4.1.3",
    "oauthlib==3.2.2",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pandas==2.1.4",
    "pandas-gbq==0.26.0",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "pipenv==2024.0.2",
    "platformdirs==4.3.3",
    "pluggy==1.5.0",
    "propcache==0.2.0",
    "proto-plus==1.23.0",
    "protobuf==4.21.12",
    "pyarrow==18.1.0",
    "pyasn1==0.6.0",
    "pyasn1_modules==0.4.0",
    "pycodestyle==2.11.1",
    "pycparser==2.22",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pydantic_core==2.14.1",
    "pydata-google-auth==1.9.0",
    "pyflakes==3.2.0",
    "PyJWT==2.10.1",
    "pyparsing==3.2.0",
    "PySocks==1.7.1",
    "pytest==8.0.0",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "python-http-client==3.3.7",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "pytz==2024.2",
    "PyYAML==6.0.1",
    "redis==5.0.1",
    "requests==2.31.0",
    "requests-oauthlib==2.0.0",
    "rsa==4.9",
    "selenium==4.28.1",
    "sendgrid==6.11.0",
    "six==1.16.0",
    "slack-logger==0.3.1",
    "slack_sdk==3.26.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "starkbank-ecdsa==2.2.0",
    "starlette==0.27.0",
    "trio==0.28.0",
    "trio-websocket==0.11.1",
    "typeguard==4.3.0",
    "typing_extensions==4.8.0",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.2.1",
    "uvicorn==0.24.0",
    "uvloop==0.22.1",
    "virtualenv==20.26.4",
    "watchfiles==1.1.1",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wsproto==1.2.0",
    "yarl==1.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]
